feat: add sync page limit cap and BigInteger progress fields
Two safety measures for sync:

1. Page limit cap (MAX_PAGES=5000) in ListenBrainz sync. Stops
   pagination when the limit is reached and sets page_limit_reached
   in the result. Pages_fetched counter survives deferral/resume.

2. BigInteger for progress_current and progress_total columns.
   Alembic migration alters Integer -> BigInteger in sync_tasks.

Fixes #12

+++ src/resonance/models/task.py
@@ -72,10 +72,10 @@ class SyncTask(base_module.Base):
         sa.Text, nullable=True, default=None
     )
     progress_current: orm.Mapped[int] = orm.mapped_column(
-        sa.Integer, nullable=False, insert_default=0
+        sa.BigInteger, nullable=False, insert_default=0
     )
     progress_total: orm.Mapped[int | None] = orm.mapped_column(
-        sa.Integer, nullable=True, default=None
+        sa.BigInteger, nullable=True, default=None
     )
     started_at: orm.Mapped[datetime.datetime | None] = orm.mapped_column(
         sa.DateTime(timezone=True), nullable=True, default=None
